# api/openapi.yaml
openapi: 3.0.4

info:
  title: Go API
  version: 1.0.0
  description: |
    Go 1.26 기반 API 템플릿
    - net/http
    - sqlc + MySQL
    - caarlos0/env
    - validator/v10
    - slog
    - OpenAPI-first + oapi-codegen

servers:
  - url: http://localhost:8080
    description: Local
  - url: https://dev-api.example.com
    description: Development
  - url: https://api.example.com
    description: Production

tags:
  - name: Health
    description: 헬스체크 API
  - name: Users
    description: 사용자 API

paths:
  /health:
    get:
      tags:
        - Health
      operationId: getHealth
      summary: 헬스체크
      description: 서버 상태를 확인한다
      responses:
        "200":
          description: 서버가 정상 상태임
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: 사용자 목록 조회
      description: 페이지네이션 기반으로 사용자 목록을 조회한다
      parameters:
        - name: page
          in: query
          required: false
          description: 페이지 번호
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          description: 페이지 크기
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 사용자 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUsersResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Users
      operationId: createUser
      summary: 사용자 생성
      description: 신규 사용자를 생성한다
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: 사용자 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 중복된 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}:
    get:
      tags:
        - Users
      operationId: getUserById
      summary: 사용자 단건 조회
      description: ID로 사용자를 조회한다
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: 사용자 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Users
      operationId: updateUserById
      summary: 사용자 수정
      description: ID로 사용자를 수정한다
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: 사용자 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 중복된 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Users
      operationId: deleteUserById
      summary: 사용자 삭제
      description: ID로 사용자를 삭제한다
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "204":
          description: 사용자 삭제 성공
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    UserIdPath:
      name: id
      in: path
      required: true
      description: 사용자 ID
      schema:
        type: integer
        format: int64
        minimum: 1

  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    CreateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: 홍길동
        email:
          type: string
          format: email
          maxLength: 255
          example: hong@example.com

    UpdateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: 홍길동
        email:
          type: string
          format: email
          maxLength: 255
          example: hong@example.com

    UserResponse:
      type: object
      required:
        - id
        - name
        - email
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 홍길동
        email:
          type: string
          format: email
          example: hong@example.com
        createdAt:
          type: string
          format: date-time
          example: "2026-03-11T15:00:00Z"

    ListUsersResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"

    PaginationResponse:
      type: object
      required:
        - page
        - size
        - total_elements
        - total_pages
        - has_next
        - has_previous
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        total_elements:
          type: integer
          format: int64
          example: 0
        total_pages:
          type: integer
          example: 0
        has_next:
          type: boolean
          example: false
        has_previous:
          type: boolean
          example: false

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: 잘못된 요청입니다
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"

    ErrorDetail:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
          example: email
        reason:
          type: string
          example: must be a valid email address